pyproject.toml
[project]
name = "proassist-rag-qa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faiss-cpu>=1.12.0",
    "hf-xet>=1.1.7",
    "huggingface-hub[hf-transfer]>=0.34.4",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "llama-cpp-python>=0.3.16",
    "pre-commit>=4.3.0",
]
